Improvements
- Cleaned up code
- Added Marks
- Implemented UISearchBar
- Removed MainView
- Merged MainViewController with RecipesTableViewController
- Added search while typing
- Improved detail appearance
