+++ Recipes/Recipes/Model/Recipe.swift
@@ -8,8 +8,6 @@
 import Foundation
 
 struct Recipe: Codable {
-    
     var name: String
     var instructions: String
-    
 }
